gym_api/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-jazzmin>=2.6.0
django-admin-interface>=0.26.0
django-colorfield>=0.11.0
psycopg2-binary>=2.9.6
gunicorn>=21.2.0
whitenoise>=6.5.0
pillow>=10.0.0
python-dateutil>=2.8.2
pytz>=2023.3
